refactor to gui_main, to avoid unnecessary runs of the same thing
