+++ src/commands/dev/mod.rs
@@ -1,3 +1,3 @@
-pub mod dev;
+pub mod dev_main;
 pub mod init_info;
 pub mod lock;
