+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![allow(unused_variables, unused_mut, dead_code)]
 pub mod commands;
+pub mod gui_main;
 pub mod gui_shell;
 pub mod keys;
 pub mod login;
